refactor: rename: common 유틸 패키지명 변경
common 유틸의 패키지명을 utils 로 변경

[skip ci]
