+++ server/edusync/src/main/java/com/codestates/edusync/model/common/utils/MemberVerificationService.java
@@ -1,4 +1,4 @@
-package com.codestates.edusync.model.common.service;
+package com.codestates.edusync.model.common.utils;
 
 import com.codestates.edusync.model.member.entity.Member;
 import org.springframework.security.core.Authentication;
