+++ server/edusync/src/main/java/com/codestates/edusync/model/common/utils/StudygroupVerifiable.java
@@ -1,4 +1,4 @@
-package com.codestates.edusync.model.common.service;
+package com.codestates.edusync.model.common.utils;
 
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
