+++ server/edusync/src/main/java/com/codestates/edusync/model/study/studygroup/service/StudygroupService.java
@@ -2,7 +2,7 @@ package com.codestates.edusync.model.study.studygroup.service;
 
 import com.codestates.edusync.exception.BusinessLogicException;
 import com.codestates.edusync.exception.ExceptionCode;
-import com.codestates.edusync.model.common.service.VerifyStudygroupUtils;
+import com.codestates.edusync.model.common.utils.VerifyStudygroupUtils;
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.model.study.studygroup.repository.StudygroupRepository;
 import com.codestates.edusync.model.study.studygroup.utils.StudygroupManager;
